nevim

+++ my_files/21. the-wild-oasis-website/app/_lib/actions.js
@@ -4,6 +4,7 @@ import { revalidatePath } from 'next/cache';
 import { auth, signIn, signOut } from './auth';
 import { supabase } from './supabase';
 import { getBookings } from './data-service';
+import { redirect } from 'next/navigation';
 
 export async function updateGuest(formData) {
   const session = await auth();
@@ -34,6 +35,10 @@ export async function deleteReservation(bookingId) {
   const guestBookings = await getBookings(session.user.guestId);
   const guestBookingsIds = guestBookings.map((booking) => booking.id);
 
+  if (guestBookingsIds.includes(bookingId)) {
+    throw new Error('you are not allowed to delete this booking');
+  }
+
   const { error } = await supabase
     .from('bookings')
     .delete()
@@ -42,6 +47,47 @@ export async function deleteReservation(bookingId) {
   if (error) throw new Error('Booking could not be deleted');
 
   revalidatePath('/account/reservations');
+}
+
+export async function updateBooking(formData) {
+  const bookingId = Number(formData.get('bookingId'));
+
+  // 1) AUTHENTICATIONS
+  const session = await auth();
+  if (!session) throw new Error('You must be logged in!');
+
+  // 2) AUTHORIZATION
+  const guestBookings = await getBookings(session.user.guestId);
+  const guestBookingsIds = guestBookings.map((booking) => booking.id);
+
+  if (!guestBookingsIds.includes(bookingId)) {
+    throw new Error('you are not allowed to update this booking');
+  }
+
+  // 3) BUILDING UPDATE DATA
+  const updateData = {
+    numGuests: Number(formData.get('numGuests')),
+    observations: formData.get('observations').slice(0, 1000),
+  };
+
+  // 4) MUTATION
+  const { error } = await supabase
+    .from('bookings')
+    .update(updateData)
+    .eq('id', bookingId)
+    .select()
+    .single();
+
+  // 5) ERROR HANDLING
+  if (error) {
+    throw new Error('Guest could not be updated');
+  }
+
+  // 6) REVALIDATION
+  revalidatePath(`account/reservations/edit/${bookingId}`);
+
+  // 7) REDIRECTING
+  redirect('/account/reservations');
 }
 
 export async function signInAction() {
